temp_requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.1
SQLAlchemy==2.0.23
Werkzeug==3.0.1
stripe==7.6.0
email-validator==2.1.0.post1
python-dotenv==1.0.0
passlib==1.7.4
PyJWT==2.8.0
requests==2.31.0
pytest==7.4.3
pytest-cov==4.1.0
Flask-Migrate==4.0.5
Flask-Caching==2.1.0
Flask-Limiter==3.5.0
Flask-Cors==4.0.0
markupsafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
Jinja2==3.1.2
